fix: show full code in IPC approval
